# 2x faster UnPrepareKey