--- a/src/Paprika/Data/NibblePath.cs
+++ b/src/Paprika/Data/NibblePath.cs
@@ -359,6 +359,7 @@ public readonly ref struct NibblePath
     /// <summary>
     /// Appends the <see cref="other1"/> and then <see cref="other2"/> path using the <paramref name="workingSet"/> as the working memory.
     /// </summary>
+    [SkipLocalsInit]
     public NibblePath Append(scoped in NibblePath other1, int hash, Span<byte> workingSet)
     {
         if (workingSet.Length <= MaxByteLength)
@@ -366,15 +367,43 @@ public readonly ref struct NibblePath
             ThrowNotEnoughMemory();
         }
 
-        // TODO: do a ref comparison with Unsafe, if the same, no need to copy!
-        WriteTo(workingSet);
+        WriteImpl(workingSet);
 
         var length = (int)Length;
         var appended = new NibblePath(ref workingSet[PreambleLength], _odd, (byte)(length + other1.Length + 2));
 
-        for (var i = 0; i < other1.Length; i++)
+        if (other1.IsEmpty == false)
         {
-            appended.UnsafeSetAt(length + i, other1[i]);
+            var alignment = (_odd + Length) ^ other1._odd;
+            if ((alignment & OddBit) == 0)
+            {
+                // oddity aligned
+                if (_odd == OddBit)
+                {
+                    // it's odd, set odd first
+                    appended.UnsafeSetAt(length, other1[0]);
+                }
+
+                // unrolled version to copy byte by byte instead of nibbles
+                var i = 0;
+                for (; i < other1.Length - 1; i+= 2)
+                {
+                    appended.GetRefAt(length + i + _odd) = other1.GetRefAt(i + _odd);
+                }
+
+                if (i < other1.Length)
+                {
+                    // it's odd, set odd first
+                    appended.UnsafeSetAt(length + other1.Length - 1, other1[other1.Length - 1]);
+                }
+            }
+            else
+            {
+                for (var i = 0; i < other1.Length; i++)
+                {
+                    appended.UnsafeSetAt(length + i, other1[i]);
+                }
+            }
         }
 
         var start = length + other1.Length;