This PR amends the `UnPrepareKey` method so that it uses a byte-aligned loop instead of the nibble aligned previous version, if the alignment allows. This reduces the time to process long keys by 50%. This should greatly reduce the time spent in preparing keys which is a hot spot for flushing down the data.

![image](https://github.com/NethermindEth/Paprika/assets/519707/aa5820f5-0b32-4ddb-8f19-f70f0445f905)


### Benchmarks

#### Before

| Method                              | sliceFrom | Mean         | Error      | StdDev     | Code Size |
|------------------------------------ |---------- |-------------:|-----------:|-----------:|----------:|
| EnumerateAll                        | ?         |  6,551.20 ns |  29.980 ns |  28.043 ns |   1,982 B |
| UnPrepareKey | 0         |    121.45 ns |   1.248 ns |   1.167 ns |   6,889 B |
| UnPrepareKey | 1         |    123.33 ns |   2.326 ns |   2.062 ns |   6,907 B |
| UnPrepareKey | 62        |     23.67 ns |   0.234 ns |   0.219 ns |   6,205 B |
| UnPrepareKey | 63        |     23.91 ns |   0.146 ns |   0.137 ns |   6,173 B |
| UnPrepareKey | 64        |     23.65 ns |   0.179 ns |   0.167 ns |   6,221 B |

#### After

| Method                              | sliceFrom | Mean         | Error      | StdDev     | Code Size |
|------------------------------------ |---------- |-------------:|-----------:|-----------:|----------:|
| EnumerateAll                        | ?         |  5,142.46 ns |  28.242 ns |  25.036 ns |   2,531 B |
| UnPrepareKey | 0         |     68.03 ns |   0.512 ns |   0.479 ns |   6,587 B |
| UnPrepareKey | 1         |     62.66 ns |   0.457 ns |   0.427 ns |   6,741 B |
| UnPrepareKey | 62        |     24.07 ns |   0.219 ns |   0.205 ns |   6,199 B |
| UnPrepareKey | 63        |     23.95 ns |   0.374 ns |   0.350 ns |   6,173 B |
| UnPrepareKey | 64        |     23.20 ns |   0.204 ns |   0.191 ns |   6,221 B |